Moving work to fragment
